% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_primers.R
\docType{methods}
\name{blastPrimers}
\alias{blastPrimers}
\alias{blastPrimers,TsIO-method}
\alias{blastPrimers,TsIOList-method}
\title{BLAST TAP-seq primers}
\usage{
blastPrimers(object, blastdb, annot, max_mismatch = 0,
  min_aligned = 0.75, tmpdir = tempdir(),
  blastn = getOption("TAPseq.blastn"))

\S4method{blastPrimers}{TsIO}(object, blastdb, annot, max_mismatch = 0,
  min_aligned = 0.75, tmpdir = tempdir(),
  blastn = getOption("TAPseq.blastn"))

\S4method{blastPrimers}{TsIOList}(object, blastdb, annot,
  max_mismatch = 0, min_aligned = 0.75, tmpdir = tempdir(),
  blastn = getOption("TAPseq.blastn"))
}
\arguments{
\item{object}{A \code{\link[TAPseq]{TsIO}} or \code{\link[TAPseq]{TsIOList}} object containing
designed primers.}

\item{blastdb}{BLAST database to which primers should be aligned. Can be created with
\code{\link[TAPseq]{createBLASTDb}}.}

\item{annot}{A \code{\link[GenomicRanges]{GRanges}} object containing exons of transcripts, which
will be used to annotate blast hits on the genome sequence. The same object as used in
\code{\link[TAPseq]{createFasta}}.}

\item{max_mismatch}{Maximum number of mismatches allowed for off-target hits (default: 0).}

\item{min_aligned}{Minimum portion of the primer sequence starting from the 3' end that must
align for off-target hits (default: 0.75).}

\item{tmpdir}{Directory needed to store temporary files.}

\item{blastn}{Path (character) to the \code{blastn} executable. Usually this is inferred when
loading/attaching the package.}
}
\value{
A \code{\link[TAPseq]{TsIO}} or \code{\link[TAPseq]{TsIOList}} object with the number
  of potential off-target priming hits added to the TAP-seq primer metadata.
}
\description{
Use BLAST to align designed TAP-seq primers against a genome and transcriptome database to
estimate off-target priming potential. Only hits where at least a specified portion of the
sequence involving the 3' end of the primer aligns with not more than a certain number of
mismatches are considered.
}
\section{Methods (by class)}{
\itemize{
\item \code{TsIO}: Blast primers in a \code{TsIO} object

\item \code{TsIOList}: Blast primers in a \code{TsIOList} object
}}

