% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bone_marrow_genex}
\alias{bone_marrow_genex}
\title{Mouse bone marrow 10x data}
\format{object of \code{Seurat} class.}
\usage{
bone_marrow_genex
}
\description{
Subset of a 10x mouse bone marrow dataset taken from Baccin et al., 2019.
(\url{https://www.nature.com/articles/s41556-019-0439-6}).
}
\keyword{datasets}
