% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_primers.R
\docType{methods}
\name{checkPrimers}
\alias{checkPrimers}
\alias{checkPrimers,TsIO-method}
\alias{checkPrimers,TsIOList-method}
\title{Check primers for complementarity}
\usage{
checkPrimers(object, primer_opt_tm = NA, primer_min_tm = NA,
  primer_max_tm = NA,
  thermo_params_path = getOption("TASCseq.thermodynamic_params_path"),
  primer3_core = getOption("TASCseq.primer3_core"))

\S4method{checkPrimers}{TsIO}(object, primer_opt_tm = NA,
  primer_min_tm = NA, primer_max_tm = NA,
  thermo_params_path = getOption("TASCseq.thermodynamic_params_path"),
  primer3_core = getOption("TASCseq.primer3_core"))

\S4method{checkPrimers}{TsIOList}(object, primer_opt_tm = NA,
  primer_min_tm = NA, primer_max_tm = NA,
  thermo_params_path = getOption("TASCseq.thermodynamic_params_path"),
  primer3_core = getOption("TASCseq.primer3_core"))
}
\arguments{
\item{object}{A \code{\link[TASCseq]{TsIO}} or \code{\link[TASCseq]{TsIOList}} object containing
designed primers.}

\item{primer_opt_tm, primer_min_tm, primer_max_tm}{Optimal, minumum and maximum primer melting
temperature.}

\item{thermo_params_path}{Path (character) to the \code{primer3_config} directory. Default set to
the same directory where \code{primer3_core} executable is found.}

\item{primer3_core}{Path (character) to the \code{primer3_core} executable. Usually this is
inferred when loading/attaching the package.}
}
\value{
A \code{\link[base]{data.frame}} with \code{check_primers} results.
}
\description{
Check a TASCseq primer set, i.e. outer or inner primers for a target gene panel, for potential
complementarity issues when multiplexing. Uses Primer3's \code{check_primers} functionality.
}
\section{Methods (by class)}{
\itemize{
\item \code{TsIO}: Check primers from \code{TsIO} objects.

\item \code{TsIOList}: Check primers from \code{TsIOList} objects.
}}

\seealso{
\url{http://primer3.org/manual.html} for Primer3 manual.
}
