% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr11_genes}
\alias{chr11_genes}
\title{Chromosome 11 genes}
\format{
object of \code{\link[GenomicRanges:GRanges-class]{GRanges}} class.
}
\usage{
chr11_genes
}
\description{
GENCODE exon annotations of all protein-coding genes within a genomic region on human chromosome
11.
}
\keyword{datasets}
