% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr11_polyA_sites}
\alias{chr11_polyA_sites}
\title{Chromosome 11 polyA sites}
\format{
object of \code{\link[GenomicRanges:GRanges-class]{GRanges}} class.
}
\usage{
chr11_polyA_sites
}
\description{
Example polyadenylation sites for expressed protein-coding genes within human chromosome 11
genomic region. This dataset was created using \code{\link[TAPseq]{inferPolyASites}} on available
K562 Drop-seq data. In a real-world application these sites would have to be pruned manually
before further use.
}
\keyword{datasets}
