% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr11_seq}
\alias{chr11_seq}
\title{Chromosome 11 sequence}
\format{object of \code{DNAStringSet} class.}
\usage{
chr11_seq
}
\description{
The human hg38 sequence of chromosome 11 as \code{\link[Biostrings]{DNAstringSet}} object. This
data is intended for examples for \code{\link[TASCseq]{getTxsSeq}}. In a real world application,
such an object would probably contain all chromosome sequences loaded from a .fasta file.
}
\keyword{datasets}
