% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr11_sequence_templates}
\alias{chr11_sequence_templates}
\title{Chromosome 11 sequence templates}
\format{object of \code{DNAStringSet} class.}
\usage{
chr11_sequence_templates
}
\description{
Sequence templates for primer design examples. The templates consist of chromosome 11 truncated
transcript sequences with the Drop-seq primer (polyA capture) added to the 3' ends.
}
\keyword{datasets}
