% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr11_truncated_txs}
\alias{chr11_truncated_txs}
\title{Chromosome 11 truncated transcripts}
\format{object of \code{GRanges} class.}
\usage{
chr11_truncated_txs
}
\description{
Annotations of transcripts within chromosome 11 region that were truncated at inferred polyA
sites using \code{\link[TAPseq]{truncateTxsPolyA}}.
}
\keyword{datasets}
