% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr11_truncated_txs_seq}
\alias{chr11_truncated_txs_seq}
\title{Chromosome 11 truncated transcript sequences}
\format{object of \code{DNAStringSet} class.}
\usage{
chr11_truncated_txs_seq
}
\description{
Sequences of truncated transcripts within human chromosome 11 region that were extracted using
\code{\link[TAPseq]{getTxsSeq}}.
}
\keyword{datasets}
