% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_primers.R
\name{createFasta}
\alias{createFasta}
\title{Create genome and transcriptome fasta file}
\usage{
createFasta(genome, output_fasta, annot = NULL, include_genome = TRUE,
  exclude_chr_pattern = "_", compress = FALSE)
}
\arguments{
\item{genome}{The \code{\link[BSgenome]{BSgenome}} object of the genome to be used to obtain the
genome and (optional) transcript sequences.}

\item{output_fasta}{Path to the output fasta file.}

\item{annot}{A \code{\link[GenomicRanges]{GRanges}} object containing all exons of transcripts to
be considered. Needs to contain \code{transcript_id} and \code{gene_name} metadata columns. To
correctly identify off-target primer binding events, \code{gene_name} should be the same as the
names of the sequence templates used for primer design! If not specified, no transcript
sequences will be included in the output fasta file.}

\item{include_genome}{(logical) Specifies whether the genome sequence should be included in the
output fasta file.}

\item{exclude_chr_pattern}{Pattern used to exclude any unwanted chromosomes from genome, such as
scaffolds with "_" by default. Set to NULL to include all chromosomes in genome.}

\item{compress}{(logical) Create a gzipped output fasta file.}
}
\description{
Create a fasta file containing the DNA sequences of a genome and its transcripts. This function
is used to create the input for \code{\link[TAPseq]{createBLASTDb}}. Currently only tested with
human and mouse genomes.
}
