% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapseq_input.R
\docType{methods}
\name{createIORecord}
\alias{createIORecord}
\alias{createIORecord,TsIO-method}
\alias{createIORecord,TsIOList-method}
\title{Create boulder IO record}
\usage{
createIORecord(object,
  thermo_params_path = getOption("TAPseq.thermodynamic_params_path"))

\S4method{createIORecord}{TsIO}(object,
  thermo_params_path = getOption("TAPseq.thermodynamic_params_path"))

\S4method{createIORecord}{TsIOList}(object,
  thermo_params_path = getOption("TAPseq.thermodynamic_params_path"))
}
\arguments{
\item{object}{TsIO of TsIOList object for which a Primer3 boulder IO record should be created.}

\item{thermo_params_path}{Path (character) to the \code{primer3_config} directory. Default set to
the same directory where \code{primer3_core} executable is found.}
}
\value{
A character vectors containing the individual lines of the IO record.
}
\description{
Takes a \code{\link[TAPseq]{TsIO}} or \code{\link[TAPseq]{TsIOList}} object and converts it
into a boulder IO record for Primer3. Essentially it converts it into a list of character vectors
that each contain the tag and the value in the form: "TAG=VALUE". More on this format can be
found in the \href{http://primer3.org/manual.html}{Primer3 manual}.
}
\details{
This function is usually not needed by the user, because functions such as
\code{\link[TAPseq]{designPrimers}} handle IO record generation. However, this function can for
instance be useful to generate IO records, write them to a file and pass them to Primer3 in the
conventional way.
}
\section{Methods (by class)}{
\itemize{
\item \code{TsIO}: Create IO record from \code{TsIO} objects.

\item \code{TsIOList}: Create IO record from \code{TsIO} objects.
}}

\examples{
library(TAPseq)

# chromosome 11 region sequence templates
data("chr11_sequence_templates")

# reverse primer used in all PCR reactions
reverse_primer <- "AAGCAGTGGTATCAACGCAGAGT"

# create TsIOList object from sequence templates
obj <- TAPseqInput(chr11_sequence_templates[1:2], reverse_primer = reverse_primer,
                   product_size_range = c(350, 500))

# create boulder IO record
boulder_io <- createIORecord(obj)
boulder_io
}
\seealso{
\url{http://primer3.org/manual.html} for Primer3 manual.
}
