% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer3_io.R
\docType{methods}
\name{designPrimers}
\alias{designPrimers}
\alias{designPrimers,TsIO-method}
\alias{designPrimers,TsIOList-method}
\title{Design primers}
\usage{
designPrimers(object, thermo_params_path = NA,
  primer3_core = getOption("TAPseq.primer3_core"))

\S4method{designPrimers}{TsIO}(object, thermo_params_path = NA,
  primer3_core = getOption("TAPseq.primer3_core"))

\S4method{designPrimers}{TsIOList}(object, thermo_params_path = NA,
  primer3_core = getOption("TAPseq.primer3_core"))
}
\arguments{
\item{object}{TsIO of TsIOList object for which primers should be designed.}

\item{thermo_params_path}{Optional path (character) to the \code{primer3_config} directory. Only
required when using Primer3 < 2.5.0.}

\item{primer3_core}{Path (character) to the \code{primer3_core} executable. Usually this is
inferred when loading/attaching the package.}
}
\value{
A new \code{TsIO} or \code{TsIOList} object containing all Primer3 output.
}
\description{
Design primers based on \code{\link[TAPseq]{TsIO}} or \code{\link[TAPseq]{TsIOList}} objects.
Creates boulder-IO records, passes input to Primer3 and parses the output.
}
\section{Methods (by class)}{
\itemize{
\item \code{TsIO}: Design primers using Primer3 from a \code{TsIO} object

\item \code{TsIOList}: Design primers using Primer3 from a \code{TsIOList} object
}}

\examples{
library(TAPseq)

# chromosome 11 region sequence templates
data("chr11_sequence_templates")

# reverse primer used in all PCR reactions
reverse_primer <- "AAGCAGTGGTATCAACGCAGAGT"

# create TsIOList object for the first two sequence templates
tapseq_io <- TAPseqInput(chr11_sequence_templates[1:2], reverse_primer = reverse_primer,
                         product_size_range = c(350, 500))

# design primers
tapseq_io <- designPrimers(tapseq_io)

# tapseq_io now contains the designed primers in the tapseq_primers slot
tapseq_primers(tapseq_io)
}
\seealso{
\url{http://primer3.org/manual.html} for Primer3 manual and \code{\link[TAPseq]{TsIO}}
  for TsIO class objects.
}
