% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{.onLoad}
\alias{.onLoad}
\title{Actions to perform when package is LOADED}
\usage{
.onLoad(libname, pkgname)
}
\description{
Check for required tools and set tool paths as package options.
}
\keyword{internal}
