% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer3_io.R
\name{parse_primer3_output}
\alias{parse_primer3_output}
\title{Parse Primer3 output into list of named character vectors containing all output per entry}
\usage{
parse_primer3_output(x)
}
\description{
Parse Primer3 output into list of named character vectors containing all output per entry
}
