% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer3_io.R
\name{parse_primer3_record}
\alias{parse_primer3_record}
\title{Parse one Primer3 output record and extract designed primers and return them as IRanges object}
\usage{
parse_primer3_record(x)
}
\description{
Parse one Primer3 output record and extract designed primers and return them as IRanges object
}
