% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_target_genes.R
\name{selectTargetGenes}
\alias{selectTargetGenes}
\title{Select target genes}
\usage{
selectTargetGenes(object, targets = NULL, expr_percentile = c(0.6,
  0.99))
}
\arguments{
\item{object}{Seurat object containing single-cell RNA-seq data from which best marker genes for
different cell populations should be learned. Needs to contain population identities for all
cell.}

\item{targets}{Desired number of target genes. At least this many target genes will be returned.
If set to NULL, the optimal number of target genes will be estimated using a cross-valdation
approach. Warning: The number of target genes might end up being very large!}

\item{expr_percentile}{Expression percentiles that candidate target genes need to fall into.
Default is 60\% to 99\%, which excludes bottom 60\% and top 1\% expressed genes from markers.}
}
\value{
A character vector containing selected target gene identifiers.
}
\description{
Select target genes that serve as markers for cell populations using a linear model with lasso
regularization.
}
\examples{
\dontrun{
library(TAPseq)
library(Seurat)

# example of mouse bone marrow 10x gene expression data
data("bone_marrow_genex")

# identify at least 200 target genes that can be used to identify cell populations
target_genes <- selectTargetGenes(bone_marrow_genex, targets = 200)

# automatically identify the number of target genes to best identify cell populations. caution:
# this can lead to very large target gene panels!
target_genes <- selectTargetGenes(bone_marrow_genex)
}
}
