% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TsIO_classes.R
\docType{methods}
\name{show,TsIO-method}
\alias{show,TsIO-method}
\title{Show method for TsIO objects}
\usage{
\S4method{show}{TsIO}(object)
}
\description{
Show method for TsIO objects
}
\keyword{internal}
