% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapseq_input.R
\name{TAPseqInput}
\alias{TAPseqInput}
\title{Create TAPseq input from sequence templates}
\usage{
TAPseqInput(sequence_templates, product_size_range,
  reverse_primer = "CTACACGACGCTCTTCCGATCT", primer_num_return = 5,
  min_primer_region = 100, primer_opt_tm = 63, primer_min_tm = 59,
  primer_max_tm = 66)
}
\arguments{
\item{sequence_templates}{A \code{\link[Biostrings]{DNAStringSet}} object containing all sequence
templates.}

\item{product_size_range}{Numerical vector of length 2 specifying the desired length of the
resulting amplicons.}

\item{reverse_primer}{Reverse primer sequence used for all PCR reactions. Default is the 10x
primer sequence: \code{CTACACGACGCTCTTCCGATCT}.}

\item{primer_num_return}{How many forward primers should be designed? (default: 5)}

\item{min_primer_region}{Minimum sequence length required for primer design. Relevant in case the
a sequence template is too short to allow the specified \code{product_size_range}.}

\item{primer_opt_tm, primer_min_tm, primer_max_tm}{Optimal, minumum and maximum primer melting
temperature. Set to NA to use Primer3s default values.}
}
\value{
\code{\link[TAPseq]{TsIOList}} object.
}
\description{
This function creates input for TAP-seq primer design from a DNAStringSet containing the
sequence templates.
}
\examples{
library(TAPseq)

# chromosome 11 region sequence templates
data("chr11_sequence_templates")

# create TsIOList object for primer desing from sequence templates
obj <- TAPseqInput(chr11_sequence_templates, product_size_range = c(350, 500))
obj
}
