% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tapseq_seq_templates.R
\name{TAPseqSeqTemplates}
\alias{TAPseqSeqTemplates}
\alias{TAPseqSeqTemplates,GRanges-method}
\alias{TAPseqSeqTemplates,GRangesList-method}
\title{Create TAPseq sequence templates}
\usage{
TAPseqSeqTemplates(transcripts, genome, beads_oligo = NA)

\S4method{TAPseqSeqTemplates}{GRanges}(transcripts, genome, beads_oligo = NA)

\S4method{TAPseqSeqTemplates}{GRangesList}(transcripts, genome, beads_oligo = NA)
}
\arguments{
\item{transcripts}{A \code{\link[GenomicRanges]{GRanges}} or
\code{\link[GenomicRanges]{GRangesList}} object containing exons of transcripts for which
sequence templates should be created. All exons in a \code{\link[GenomicRanges]{GRanges}}
object are assumed to belong to the same transcript. Multiple transcripts can be provided in a
\code{\link[GenomicRanges]{GRangesList}} object.}

\item{genome}{A \code{\link[BSgenome]{BSgenome}} or \code{\link[Biostrings]{DNAStringSet}}
object containing chromosome sequences which should be used to extract transcript sequences.}

\item{beads_oligo}{(character) The Beads-oligo-dT sequence used for droplet sequencing. If
nothing is specified (\code{beads_oligo = NA}), the 10x V3 Beads-oligo-dT sequence is used.
Can be changed if primers are for instance designed for Drop-seq. Any barcode bases need to be
replaced by \code{N}.}
}
\value{
A \code{\link[Biostrings]{DNAString}} or \code{\link[Biostrings]{DNAStringSet}} object
  containing the sequence template(s).
}
\description{
Methods to create sequence templates for TAP-seq primer design. This extracts the transcript
sequence from the provided genome and adds the reverse complement of the "Beads-oligo-dT"
sequence to the 3' end.
}
\section{Methods (by class)}{
\itemize{
\item \code{GRanges}: Create sequence templates from \code{GRanges} input

\item \code{GRangesList}: Create sequence templates from \code{GRangesList} input
}}

\examples{
\dontrun{
library(TAPseq)
library(Biostrings)
library(BSgenome)

# truncated transcripts for chr11 target genes
data("chr11_truncated_txs")

# human genome (hg38) BSgenome object (needs to be istalled separately from Bioconductor)
# the genome sequence could also be loaded from a .fasta file with
hg38 <- getBSgenome("BSgenome.Hsapiens.UCSC.hg38")

# create sequence templates for all target transcripts on chr11
seq_templates <- TAPseqSeqTemplates(chr11_truncated_txs, genome = hg38)

# create sequence templates using the Drop-seq Beads-oligo-dT sequence
ds_oligo <- "TTTTTTTAAGCAGTGGTATCAACGCAGAGTACNNNNNNNNNNNNNNNNNNNNTTTTTTTTTTTTTTTTTTTTTTTTTTTTTT"
seq_templates <- TAPseqSeqTemplates(chr11_truncated_txs, genome = hg38, beads_oligo = ds_oligo)
}
}
