% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tascseq_input.R
\name{TASCseqInput}
\alias{TASCseqInput}
\title{Create TASCseq input from sequence templates}
\usage{
TASCseqInput(sequence_templates, reverse_primer, product_size_range,
  primer_num_return = 5, min_primer_region = 100, primer_opt_tm = NA,
  primer_min_tm = NA, primer_max_tm = NA)
}
\arguments{
\item{sequence_templates}{A \code{\link{DNAStringSet}} object containing all sequence templates.}

\item{reverse_primer}{Reverse primer sequence used for all PCR reactions.}

\item{product_size_range}{Numerical vector of length 2 specifying the desired length of the
resulting amplicons.}

\item{primer_num_return}{How many forward primers should be designed? (default: 5)}

\item{min_primer_region}{Minimum sequence length required for primer design. Relevant in case the
a sequence template is too short to allow the specified \code{product_size_range}.}

\item{primer_opt_tm, primer_min_tm, primer_max_tm}{Optimal, minumum and maximum primer melting
temperature.}
}
\value{
\code{\link[TASCseq]{TsIOList}} object.
}
\description{
This function creates input for TASC-seq primer design from a DNAStringSet containing the
sequence templates.
}
