% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TsIO_classes.R
\name{accessors}
\alias{accessors}
\alias{sequence_id}
\alias{sequence_id<-}
\alias{sequence_template}
\alias{sequence_template<-}
\alias{reverse_primer}
\alias{reverse_primer<-}
\alias{product_size_range}
\alias{product_size_range<-}
\alias{primer_num_return}
\alias{primer_num_return<-}
\alias{min_primer_region}
\alias{min_primer_region<-}
\alias{primer_opt_tm}
\alias{primer_opt_tm<-}
\alias{primer_min_tm}
\alias{primer_min_tm<-}
\alias{primer_max_tm}
\alias{primer_max_tm<-}
\alias{tascseq_primers}
\alias{pcr_products}
\title{Accessors for TsIO objects}
\usage{
sequence_id(x)

sequence_id(x) <- value

sequence_template(x)

sequence_template(x) <- value

reverse_primer(x)

reverse_primer(x) <- value

product_size_range(x)

product_size_range(x) <- value

primer_num_return(x)

primer_num_return(x) <- value

min_primer_region(x)

min_primer_region(x) <- value

primer_opt_tm(x)

primer_opt_tm(x) <- value

primer_min_tm(x)

primer_min_tm(x) <- value

primer_max_tm(x)

primer_max_tm(x) <- value

tascseq_primers(x)

pcr_products(x)
}
\arguments{
\item{x}{A \code{TsIO} or \code{TsIOList} class object.}

\item{value}{A valid value to assign to the chosen slot.}
}
\description{
A set of functions for getting/setting/modifying the data stored in \code{\link[TASCseq]{TsIO}}
or \code{\link[TASCseq]{TsIOList}} class objects.
}
