% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{check_tool_installation}
\alias{check_tool_installation}
\title{Check required tools}
\usage{
check_tool_installation()
}
\value{
A \code{list} containing paths to required tools, if found, else NA.
}
\description{
Check if required software is installed and return paths to executables (if found).
}
\keyword{internal}
