% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast_primers.R
\name{createBLASTDb}
\alias{createBLASTDb}
\title{Create BLAST database}
\usage{
createBLASTDb(blastdb, fasta,
  makeblastdb = getOption("TASCseq.makeblastdb"),
  compression = c("auto", "gzip", "none"), title = "None")
}
\arguments{
\item{blastdb}{A file path to the blastdb that should be created.}

\item{fasta}{A fasta file containing all sequences to be included in the database. Can be created
with \code{\link[TASCseq]{createFasta}}.}

\item{makeblastdb}{Path (character) to the \code{makeblastdb} executable. Usually this is
inferred when loading/attaching the package.}

\item{compression}{What compression was used when generating the fasta file? Default is "auto",
which tries to infer compression from fasta filename. Currently only allows for gzip
compression (.gz file extension).}

\item{title}{Optional title for BLAST database.}
}
\description{
Create a database which can be used to estimate potential off-target priming of TASC-seq primers.
}
