% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tascseq_input.R
\docType{methods}
\name{createIORecord}
\alias{createIORecord}
\alias{createIORecord,TsIO-method}
\alias{createIORecord,TsIOList-method}
\title{Create boulder IO record}
\usage{
createIORecord(object,
  thermo_params_path = getOption("TASCseq.thermodynamic_params_path"))

\S4method{createIORecord}{TsIO}(object,
  thermo_params_path = getOption("TASCseq.thermodynamic_params_path"))

\S4method{createIORecord}{TsIOList}(object,
  thermo_params_path = getOption("TASCseq.thermodynamic_params_path"))
}
\arguments{
\item{object}{TsIO of TsIOList object for which a Primer3 boulder IO record should be created.}

\item{thermo_params_path}{Path (character) to the \code{primer3_config} directory. Default set to
the same directory where \code{primer3_core} executable is found.}
}
\value{
A character vectors containing the individual lines of the IO record.
}
\description{
Takes a \code{\link[TASCseq]{TsIO}} or \code{\link[TASCseq]{TsIOList}} object and converts it
into a boulder IO record for Primer3. Essentially it converts it into a list of character vectors
that each contain the tag and the value in the form: "TAG=VALUE". More on this format can be
found in the \href{http://primer3.org/manual.html}{Primer3 manual}.
}
\details{
This function is usually not needed by the user, because functions such as
\code{\link[TASCseq]{designPrimers}} handle IO record generation. However, this function can for
instance be useful to generate IO records, write them to a file and pass them to Primer3 in the
conventional way.
}
\section{Methods (by class)}{
\itemize{
\item \code{TsIO}: Create IO record from \code{TsIO} objects.

\item \code{TsIOList}: Create IO record from \code{TsIO} objects.
}}

\seealso{
\url{http://primer3.org/manual.html} for Primer3 manual.
}
