% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer3_io.R
\docType{methods}
\name{designPrimers}
\alias{designPrimers}
\alias{designPrimers,TsIO-method}
\alias{designPrimers,TsIOList-method}
\title{Design primers}
\usage{
designPrimers(object,
  thermo_params_path = getOption("TAPseq.thermodynamic_params_path"),
  primer3_core = getOption("TAPseq.primer3_core"))

\S4method{designPrimers}{TsIO}(object,
  thermo_params_path = getOption("TAPseq.thermodynamic_params_path"),
  primer3_core = getOption("TAPseq.primer3_core"))

\S4method{designPrimers}{TsIOList}(object,
  thermo_params_path = getOption("TAPseq.thermodynamic_params_path"),
  primer3_core = getOption("TAPseq.primer3_core"))
}
\arguments{
\item{object}{TsIO of TsIOList object for which primers should be designed.}

\item{thermo_params_path}{Path (character) to the \code{primer3_config} directory. Default set to
the same directory where \code{primer3_core} executable is found.}

\item{primer3_core}{Path (character) to the \code{primer3_core} executable. Usually this is
inferred when loading/attaching the package.}
}
\value{
A new \code{TsIO} or \code{TsIOList} object containing all Primer3 output.
}
\description{
Design primers based on \code{\link[TAPseq]{TsIO}} or \code{\link[TAPseq]{TsIOList}} objects.
Creates boulder-IO records, passes input to Primer3 and parses the output.
}
\section{Methods (by class)}{
\itemize{
\item \code{TsIO}: Design primers using Primer3 from a \code{TsIO} object

\item \code{TsIOList}: Design primers using Primer3 from a \code{TsIOList} object
}}

\seealso{
\url{http://primer3.org/manual.html} for Primer3 manual.
}
