% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primer3_io.R
\name{parsePrimer3Output}
\alias{parsePrimer3Output}
\title{Parse Primer3 Output}
\usage{
parsePrimer3Output(object, primer3_output)
}
\arguments{
\item{object}{The \code{\link[TAPseq]{TsIO}} or \code{\link[TAPseq]{TsIOList}} object used to
design primers. No errors or warnings if this is another \code{TsIO} or \code{TsIOList} object!}

\item{primer3_output}{Character vector containing raw Primer3 output.}
}
\value{
\code{TsIO} or \code{TsIOList} object with added Primer3 output
}
\description{
Parse Primer3 output and add to input \code{\link[TAPseq]{TsIO}} or
\code{\link[TAPseq]{TsIOList}} object.
}
\keyword{internal}
