% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_target_genes.R
\name{plotTargetGenes}
\alias{plotTargetGenes}
\title{Plot target genes set}
\usage{
plotTargetGenes(object, target_genes, npcs = 15)
}
\arguments{
\item{object}{Seurat object containing single-cell RNA-seq data for all genes. Needs to contain
population identities for all cell.}

\item{target_genes}{(character) Target gene names.}

\item{npcs}{(integer) Number of principal components to use for UMAP.}
}
\description{
Plot how well a selected set of target genes discriminates between cell populations using UMAP
visualization.
}
