% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TsIO_classes.R
\docType{class}
\name{TsIOList-class}
\alias{TsIOList-class}
\alias{TsIOList}
\alias{sequence_template,TsIOList-method}
\alias{tascseq_primers,TsIOList-method}
\alias{pcr_products,TsIOList-method}
\title{TsIOList class}
\usage{
TsIOList(...)

\S4method{sequence_template}{TsIOList}(x)

\S4method{tascseq_primers}{TsIOList}(x)

\S4method{pcr_products}{TsIOList}(x)
}
\arguments{
\item{...}{Multiple TsIO objects from which a TsIOList object should be created.}
}
\value{
A \code{TsIOList} object.
}
\description{
TsIOList class is a container to store multiple \code{\link[TASCseq]{TsIO}} objects. This enables
storing of Primer3 input and output for multiple target genes.
}
\section{Methods (by generic)}{
\itemize{
\item \code{sequence_template}: Get sequence_templates

\item \code{tascseq_primers}: Get tascseq_primers

\item \code{pcr_products}: Get pcr_products
}}

\seealso{
\link[TASCseq]{TsIO}
}
